Matsumoto: Scopes accepts Float numbers as declared effort
git-svn-id: http://192.168.2.95/svn/repos/trunk@747 014e18b7-1377-44f3-b398-c31593b6cf9d

+++ ontrack/src/main/java/br/com/oncast/ontrack/client/ui/components/scopetree/actions/internal/DeclareEffortInternalAction.java
@@ -31,11 +31,11 @@ public class DeclareEffortInternalAction implements InternalAction {
 
 	@Override
 	public ModelAction createEquivalentModelAction(final String value) {
-		int declaredEffort;
+		float declaredEffort;
 		boolean hasDeclaredEffort;
 
 		try {
-			declaredEffort = Integer.valueOf(value);
+			declaredEffort = Float.valueOf(value);
 			hasDeclaredEffort = (value != null && !value.isEmpty());
 		}
 		catch (final NumberFormatException e) {
